Encapsulate sending GET requests
UrlGet is the best name I could think of for now. Still don't know how I
feel about objects with class methods only or creating new instances for
these objects(and making instance methods). So far we've been going with
the class method objects, as every object we've built so far only has one
(class)method. This is the first object with more than one method, but
the main reason for pulling headers into a method is to assert the
request was made using them. We still need to ensure we follow redirects
and handle non 200 response codes, but at this point we aren't even using
UrlGet yet.

We plan to design the class(or classes) which will orchestrate using all the
objects being created to be implemented last. This might be a terrible
idea, but for now it seems to make sense to continue building pieces one
at a time.

After typing this all out I think this exact scenario was mentioned in
growing object-oriented software guided by tests as a bad idea because
of how hard it was to integrate the pieces after the fact. Hopefully
this isn't a bad idea...