initial implementation for encapsulating POST
We might create separate classes for each POST request we are going to
make. For now, a lot of the implementation details are the same for both
POST requests we plan to make. As Sandi Metz says: "duplication is far
cheaper than the wrong abstraction".

We also include the dotenv gem, allowing us to hide some of the
sensitive information in this project. The logic is that if a
.env.production file exists, its values will be used instead of the
checked-in .env.development file.