Add to cart
We've decided that UrlPost should only be concerned with encapsulating
sending POST requests. These changes move the headers for adding to cart
to their own class and restore UrlPost to being concerned with only
sending post requests(we left the checkout headers in UrlPost because we
have not yet implemented any checkout logic).