Checkout object
These changes remove the remaining endpoint specific logic from UrlPost
into its own object.

We originally had Checkout.url but it seems something with RestClient
doesn't play well with a class method named url without any parameters.
We also rename AddToCart.url even though it is not affected by this.

Finally, we still have the hash to html params string problem as in
4b89f1a124ebd331644b1e4a6dafb89c874ee416. We will probably add
ActiveSupport when we need to start asserting the params in Checkout are
correct.