no longer require_relative
At first we did not want to deal with the directory structure and it was
easy to just require_relative to get our tests passing. Now that we have
several classes and plan to add more, we need to get a hold of this
requiring files business. These changes make it so any new class added
under lib gets added to lib/dogedeini.rb and we no longer need to
require the file for the class under test in the test.